Remove "equally"
From Dave Guarino: This will be unexpected, but I vote to remove equally! I think I understand the motivation, but I'd rather say "serve everyone with dignity." My reason? I've actually seen the "equal" discourse engaged really concretely to PREVENT any change, and in particular small-scale experimentation.  More broadly, my disagreement with this is grounded in two points: (1) the status quo is already unequal; and (2) any change to the status quo also has distributional effects, even if it makes things way better for 99.9% of people but worse for 0.1%. I think if we value experimentation and rapid learning to improve the system for all we need to accept that there will be intermediate distributional effects. [Note: if you are still reading by this point let's have coffee some time.] The best approach is not to hold hard and fast to equality (which is illusory) but engage in a more intentional approach: if you're making it better for some subset of users, consider making it better for a subset already disproportionately underserved by the system. But anchoring to "equality" writ large as a high-level goal prevents us from most effectively making it better for people.